requirements.txt
mcp
wordlift-client>=1.123.0
rdflib
python-dotenv
